Got Recorded Autonomous working, implemented Encoders into RecordedA utonomous to make the driving more accurate (untested), renamed the RecordedAutonomous classes to Autonomous, renamed CanGrabber classes to BinGrabber, and did some various little changes.

+++ 2015 Robot/src/org/usfirst/frc/team1902/robot/commands/LiftCommand.java
@@ -1,14 +1,13 @@
 package org.usfirst.frc.team1902.robot.commands;
 
 import org.usfirst.frc.team1902.robot.Robot;
-
 import edu.wpi.first.wpilibj.command.Command;
 
 public class LiftCommand extends Command {
 
-	public int motorValue;
+	public double motorValue;
 	
-    public LiftCommand(int motorValue) {
+    public LiftCommand(double motorValue) {
         requires(Robot.lift);
         this.motorValue = motorValue;
     }
